refactor: Container를 DetailInput으로 변경
